+++ bin/workflow_glue/report_utils/common.py
@@ -10,6 +10,8 @@ CHROMOSOMES = {str(i): i for i in range(1, 23)}
 CHROMOSOMES.update({f"chr{i}": int(i) for i in CHROMOSOMES})
 CHROMOSOMES.update({'X': 23, 'Y': 24, 'chrX': 23, 'chrY': 24})
 THEME = "epi2melabs"
+CLINVAR_BASE = "https://www.ncbi.nlm.nih.gov/clinvar/variation/"
+NCBI_GENE_BASE = "https://www.ncbi.nlm.nih.gov/gene/"
 
 
 def hist_plot(